requirements.txt
web3
python-dotenv
requests
pyyaml